backend/requirements.txt
setuptools
aiosqlite==0.20.0
fastapi==0.111.0
httpx==0.27.0
numpy==1.26.4
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
uvicorn[standard]==0.29.0
websockets==12.0
